added some icons to see the current state of the task, made a task icon script to show the correct icon based on the current task state, added task events for starting, advancing and completing a task, created a manager to keep track of tasks and check task requirements, created a taskpoint script to handle task start and end points,

+++ Assets/Scripts/Rewards/RewardEvents.cs
@@ -26,7 +26,7 @@ public class RewardEvents
 
     public event Action<float> onTimeGained;
 
-    public void TimeGained(int time)
+    public void TimeGained(float time)
     {
         if (onTimeGained != null)
         {
